% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinywqbench-package.R
\docType{package}
\name{shinywqbench-package}
\alias{shinywqbench}
\alias{shinywqbench-package}
\title{shinywqbench: R Shiny App to Generate Water Quality Guidelines Benchmarks}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})

Other contributors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID}) [contributor]
}

}
\keyword{internal}
